Got the database working

+++ src/main/java/my/gcu/data/entity/repository/ProductRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import my.gcu.data.entity.ProductEntity;
 
-public interface ProductRepository extends CrudRepository<ProductEntity, Long>
+public interface ProductRepository extends CrudRepository<ProductEntity, Integer>
 {
 
 }
